WIP: Refactor and make way for yarn workspaces
