+++ src/Esy.re
@@ -1,3 +1,4 @@
+/** Esy API **/
 type status = {
   isProject: bool,
   isProjectSolved: bool,
@@ -99,110 +100,5 @@ let getStatus = path => {
            reject(Stderr(statusErrorString));
          }
        )
-  );
-};
-
-let dropAnEsyJSON = (~compilerVersion, ~folder) => {
-  let esyJsonTargetDir = Path.join([|folder, ".vscode", "esy"|]);
-  Js.Promise.(
-    Fs.mkdir(~p=true, esyJsonTargetDir)
-    |> then_(() =>
-         Fs.writeFile(
-           Filename.concat(esyJsonTargetDir, "esy.json"),
-           thisProjectsEsyJson,
-         )
-       )
-  );
-};
-
-let processDeps = (dependenciesJson, folder) => {
-  Js.Promise.(
-    Js.Json.(
-      switch (dependenciesJson->object_->classify) {
-      | JSONObject(dependenciesDict) =>
-        switch (Js.Dict.get(dependenciesDict, "bs-platform")) {
-        | Some(bsPlatformVersionJson) =>
-          switch (Js.Json.classify(bsPlatformVersionJson)) {
-          | JSONString(bsPlatformVersion) =>
-            if (bsPlatformVersion
-                ->Semver.minVersion
-                ->Semver.satisfies(">=6.0.0")) {
-              dropAnEsyJSON(~folder, ~compilerVersion="4.6.x");
-            } else {
-              dropAnEsyJSON(~folder, ~compilerVersion="4.2.x");
-            }
-          | _ =>
-            reject(
-              Failure(
-                "'bs-platform' (in dependencies section) was expected to contain a semver string, but it was not!",
-              ),
-            )
-          }
-
-        | None =>
-          reject(
-            Failure(
-              "'bs-platform' was expected in the 'dependencies' section of the manifest file, but was not found!",
-            ),
-          )
-        }
-      | _ =>
-        reject(
-          Failure(
-            "'dependencies' section in the manifest file was expected to be dictionary, but it was not!",
-          ),
-        )
-      }
-    )
-  );
-};
-
-let getSubDict = (dict, key) =>
-  dict->Js.Dict.get(key)->Belt.Option.flatMap(Js.Json.decodeObject);
-
-let mergeDicts = (dict1, dict2) =>
-  Js.Dict.fromArray(
-    Js.Array.concat(Js.Dict.entries(dict1), Js.Dict.entries(dict2)),
-  );
-
-let setup = (~manifestPath) => {
-  Js.Promise.(
-    Js.Json.(
-      {
-        let folder = Filename.dirname(manifestPath);
-        Fs.readFile(manifestPath)
-        |> then_(manifest => {
-             let manifestJson = parseExn(manifest);
-             switch (classify(manifestJson)) {
-             | JSONObject(dict) =>
-               switch (
-                 getSubDict(dict, "dependencies"),
-                 getSubDict(dict, "devDependencies"),
-               ) {
-               | (Some(dependenciesJson), None)
-               | (None, Some(dependenciesJson)) =>
-                 processDeps(dependenciesJson, folder)
-               | (Some(dependenciesJson), Some(devDependenciesJson)) =>
-                 processDeps(
-                   mergeDicts(dependenciesJson, devDependenciesJson),
-                   folder,
-                 )
-               | (None, None) =>
-                 reject(
-                   Failure(
-                     "The manifest file doesn't seem to contain `dependencies` or `devDependencies` property",
-                   ),
-                 )
-               }
-             | _ =>
-               reject(
-                 Failure(
-                   "The entire manifest was expected to be dictionary of key-vals, but it was not!:",
-                 ),
-               )
-             };
-           });
-      }
-    )
   );
 };
