+++ src/Utils.re
@@ -20,3 +20,11 @@ let mapResultAndResolvePromise = (f, r) =>
 
 let bindResultAndResolvePromise = (f, r) =>
   Js.Promise.(Result.(r >>= f |> resolve));
+
+let getSubDict = (dict, key) =>
+  dict->Js.Dict.get(key)->Belt.Option.flatMap(Js.Json.decodeObject);
+
+let mergeDicts = (dict1, dict2) =>
+  Js.Dict.fromArray(
+    Js.Array.concat(Js.Dict.entries(dict1), Js.Dict.entries(dict2)),
+  );
